Proof of Concept: Filter all RTP packets through GStreamer
Packets are passed to GStreamer via the GstAppSrc element, which gets
pushed with new buffers.

Once inside the GStreamer Pipeline, we can do anything as desired;
currently, an "identity" element is used to simulate a filter.

Lastly, packets are retrieved back from GStreamer by pulling from the
GstAppSink. Then, a VERY UGLY thing is done: the GstBuffer is memcopied
back directly to the RtpPacket memory. Whatch out because this is asking
for segfaults!! But for now it works because the "identity" element
doesn't change the data at all.